Todo 006 users can login and crud their own lists

+++ app/controllers/tasks_controller.rb
@@ -1,30 +1,47 @@
 class TasksController < ApplicationController
 
+    include TasksHelper
     rescue_from ::ActiveRecord::RecordNotFound, with: :record_not_found
 
     def create
-        task = Task.new(task_params)
-        if task.save
-            render_response(task)
+        if current_user
+            task = current_user.tasks.new(description: task_params[:description])
+            if task.save
+                render_response("The task was successfully created.")
+            else
+                render_response("The task couldn't be added. Try again!")
+            end
         else
-            render_response("The task couldn't be added. Try again!")
+            render_response("You are not authorized to add this task.")
         end
     end
 
     def index
-        render_response(Task.all)
+        if current_user
+            render_response(current_user.tasks)
+        else
+            render_response("You are not authorized to see this list.")
+        end
     end
 
     def update
         task = Task.find(params[:id])
-        task.update(is_completed: params[:is_completed])
-        render json: "The task was successfully updated."
+        if task_belons_to_user?(task.user_id)
+            task.update(is_completed: params[:is_completed])
+            render_response("The task was successfully updated.")
+        else
+            render_response("You are not authorized to update this task.")
+        end
     end
 
     def destroy
         task = Task.find(params[:id])
-        task.destroy
-        render_response("The task was successfully deleted.")
+        if task_belons_to_user?(task.user_id)
+            task.destroy
+            render_response("The task was successfully deleted.")
+        else
+            render_response("You are not authorized to delete this task.")
+        end
     end
 
     private
